Completed functions to choose player controls
Finished functions in the PongUtil class to allow player to choose controls. The chooseControls function prompts user for the left/right players' up and down keys (making use of the chooseKey function), the confirmOption function allows the user to change their mind. The checkKeyMappings function ensures that no key matches another key and will prompt the user to enter a valid option. Once the chooseControls function finishes executing, it returns a char[] array. Will have to implement a fix to allow user to enter up and down keys for instance.